work for the https://github.com/surveyjs/survey-creator/issues/2065
