+++ packages/survey-creator-core/src/components/item-value.ts
@@ -11,6 +11,8 @@ import { DragDropChoices } from "survey-core";
 import "./item-value.scss";
 import { getLocString } from "../editorLocalization";
 
+import {DragOrClickHelper} from "../utils/dragOrClickHelper";
+
 export class ItemValueWrapperViewModel extends Base {
   @property({ defaultValue: false }) isNew: boolean;
   @property({ defaultValue: false }) isDragging: boolean;
@@ -38,50 +40,18 @@ export class ItemValueWrapperViewModel extends Base {
     this.dragDropHelper.onGhostPositionChanged.add(
       this.handleDragDropGhostPositionChanged
     );
+    this.dragOrClickHelper = new DragOrClickHelper(this.startDragItemValue);
   }
 
-  // correct handle click vs drag
-  private pointerDownEvent;
-  private currentTarget;
-  private startX;
-  private startY;
-  private currentX;
-  private currentY;
+  private dragOrClickHelper: DragOrClickHelper;
 
   onPointerDown(pointerDownEvent: PointerEvent) {
-    this.pointerDownEvent = pointerDownEvent;
-    this.currentTarget = pointerDownEvent.currentTarget;
-    this.startX = pointerDownEvent.pageX;
-    this.startY = pointerDownEvent.pageY;
-    document.addEventListener("pointermove", this.startDragItemValue);
-    this.currentTarget.addEventListener("pointerup", this.onPointerUp);
+    this.dragOrClickHelper.onPointerDown(pointerDownEvent);
   }
 
-  startDragItemValue = (pointerMoveEvent: PointerEvent) => {
-    this.currentX = pointerMoveEvent.pageX;
-    this.currentY = pointerMoveEvent.pageY;
-    if (this.isMicroMovement) return;
-    this.clearListeners();
-
-    const event = this.pointerDownEvent;
-    this.dragDropHelper.startDrag(event, this.item, this.question, <HTMLElement>this.currentTarget);
+  startDragItemValue = (pointerDownEvent: PointerEvent, currentTarget: any) => {
+    this.dragDropHelper.startDrag(pointerDownEvent, this.item, this.question, <HTMLElement>currentTarget);
   }
-  public onPointerUp = (event) => {
-    this.clearListeners();
-  };
-  // see https://stackoverflow.com/questions/6042202/how-to-distinguish-mouse-click-and-drag
-  private get isMicroMovement() {
-    const delta = 10;
-    const diffX = Math.abs(this.currentX - this.startX);
-    const diffY = Math.abs(this.currentY - this.startY);
-    return diffX < delta && diffY < delta;
-  }
-  private clearListeners() {
-    if (!this.pointerDownEvent) return;
-    document.removeEventListener("pointermove", this.startDragItemValue);
-    this.currentTarget.removeEventListener("pointerup", this.onPointerUp);
-  }
-  // EO correct handle click vs drag
 
   private get dragDropHelper(): DragDropChoices {
     return this.creator.dragDropChoices;
